Bugs

+++ app/controllers/students/quiz_requests_controller.rb
@@ -2,8 +2,10 @@
 module Students
   class QuizRequestsController < BaseController
     def destroy
-      request = current_user.quiz_lock || QuizRequest.find(params[:id])
-      request.destroy
+      request = current_user.quiz_lock || QuizRequest.find_by_id(params[:id])
+      if request
+        request.destroy
+      end
       flash[:success] = 'Cancelled quiz request!'
       redirect_to students_dashboard_path
     end
